external: uvc-gadget: make event dispath safe
Using list_for_each_entry_safe() means that it would be safe
to watch/unwatch the current file descriptor being processed,
as it caches the next element to be processed in a local register.

But things can break if watch/unwatch file descriptor that is NOT
being processed in the event list.

To overcome the limitation, here introduce a simple event iterator
that caches the events needed to process from the event list, and
leave the event list to a "free" state and thus add / delete
elements at any context are permited.